Renamed TaskSystem to ScheduleBasedExecution to avoid confusion, since other systems also can use tasks
