+++ Assets/Virtual Agents Framework/Runtime/Scripts/BaseTask.cs
